tests/wpt/web-platform-tests/payment-request/payment-request-multiple-show-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Manual test for multiple PaymentRequest.show()</title>
<link rel="help" href="https://w3c.github.io/payment-request/#show-method">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
"use strict";
setup({
  allow_uncaught_exception: true,
});
const defaultMethods = Object.freeze([
  { supportedMethods: "basic-card" },
  {
    supportedMethods: "https://apple.com/apple-pay",
    data: {
      version: 3,
      merchantIdentifier: "merchant.com.example",
      countryCode: "US",
      merchantCapabilities: ["supports3DS"],
      supportedNetworks: ["visa"],
    }
  },
]);

const defaultDetails = Object.freeze({
  total: {
    label: "Total",
    amount: {
      currency: "USD",
      value: "1.00",
    },
  },
});

function testCallingShowMultipleTimes() {
  promise_test(async t => {
    const request = new PaymentRequest(defaultMethods, defaultDetails);
    const p1 = request.show();
    const p2 = request.show();
    const p3 = request.show();
    const promises = new Set([p1, p2, p3]);
    await request.abort();
    assert_equals(promises.size, 3, "Must have three unique objects");
    await promise_rejects(t, "AbortError", p1);
    await promise_rejects(t, "InvalidStateError", p2);
    await promise_rejects(t, "InvalidStateError", p3);
  }, "Calling show() multiple times is always a new object.");
}
</script>
<h2>Manual test for multiple PaymentRequest.show()</h2>
<p>
  Click on the button to bring up the Payment Request UI window and then will
  close it automatically. (If a payment sheet stays open, the test has failed.)
</p>
<ul>
  <li>
    <button onclick="testCallingShowMultipleTimes()">
      Calling show() multiple times is always a new object.
    </button>
  </li>
</ul>
<small>
  If you find a buggy test, please <a href="https://github.com/web-platform-tests/wpt/issues">file a bug</a>
  and tag one of the <a href="https://github.com/web-platform-tests/wpt/blob/master/payment-request/META.yml">suggested reviewers</a>.
</small>
